+++ src/sawe/project.h
@@ -6,13 +6,16 @@
 #include "tools/toolfactory.h"
 #include "tools/commands/commandinvoker.h"
 #include "signal/processing/chain.h"
+#include "sawedll.h"
 
 // boost
 #include <boost/scoped_ptr.hpp>
 #include <boost/serialization/version.hpp>
 #include <boost/serialization/split_member.hpp>
+#include <boost/serialization/extended_type_info.hpp>
+#include <boost/serialization/singleton.hpp>
 #include <boost/serialization/shared_ptr.hpp>
-#include <boost/serialization/vector.hpp> 
+#include <boost/serialization/vector.hpp>
 #include <boost/serialization/binary_object.hpp> 
 #include <boost/serialization/version.hpp>
 
@@ -167,7 +170,7 @@ public:
     Signal::Processing::Chain::ptr processing_chain() { return processing_chain_; }
     Signal::Processing::TargetMarker::ptr default_target();
     Signal::OperationDesc::Extent extent();
-    float length();
+    double length();
 
 private:
     Project(); // used by deserialization
